+++ packages/message-bus-adapter-event-bridge-s3/.lintstagedrc.js
