+++ packages/message-bus-adapter-event-bridge-s3/README.md
@@ -2,7 +2,7 @@
 
 DRY Castore [`MessageBus`](https://github.com/castore-dev/castore/#--messagebus) definition using [AWS EventBridge](https://aws.amazon.com/eventbridge/) and [AWS S3](https://aws.amazon.com/s3/).
 
-This adapter works like the [EventBridge Message Bus Adapter](https://www.npmjs.com/package/@castore/event-bridge-message-bus-adapter) (it actually uses it under the hood), excepts that entry sizes are checked before publishing messages to EventBridge. If they are over the [256KB limit](https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-putevent-size.html), they are written on a s3 bucket instead, and a message is sent containing a pre-signed URL, as [recommended by AWS](https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-putevent-size.html).
+This adapter works like the [EventBridge Message Bus Adapter](https://www.npmjs.com/package/@castore/message-bus-adapter-event-bridge) (it actually uses it under the hood), excepts that entry sizes are checked before publishing messages to EventBridge. If they are over the [256KB limit](https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-putevent-size.html), they are written on a s3 bucket instead, and a message is sent containing a pre-signed URL, as [recommended by AWS](https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-putevent-size.html).
 
 Do not forget to set a [lifecycle configuration](https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-lifecycle-mgmt.html) on your s3 bucket to delete the written objects after the presigned URL has expired to avoid high s3 bills! 🤑
 
@@ -10,10 +10,10 @@ Do not forget to set a [lifecycle configuration](https://docs.aws.amazon.com/Ama
 
 ```bash
 # npm
-npm install @castore/event-bridge-s3-message-bus-adapter
+npm install @castore/message-bus-adapter-event-bridge-s3
 
 # yarn
-yarn add @castore/event-bridge-s3-message-bus-adapter
+yarn add @castore/message-bus-adapter-event-bridge-s3
 ```
 
 This package has `@castore/core`, `@aws-sdk/client-eventbridge` (above v3), `@aws-sdk/client-eventbridge` (above v3) and `@aws-sdk/s3-request-presigner` (above v3) as peer dependencies, so you will have to install them as well:
@@ -32,7 +32,7 @@ yarn add @castore/core @aws-sdk/client-eventbridge @aws-sdk/client-s3 @aws-sdk/s
 import { EventBridgeClient } from '@aws-sdk/client-eventbridge';
 import { S3Client } from '@aws-sdk/client-s3';
 
-import { EventBridgeS3MessageBusAdapter } from '@castore/event-bridge-s3-message-bus-adapter';
+import { EventBridgeS3MessageBusAdapter } from '@castore/message-bus-adapter-event-bridge-s3';
 
 const eventBridgeClient = new EventBridgeClient({});
 const s3Client = new S3Client({});
@@ -65,7 +65,7 @@ This will directly plug your MessageBus to EventBridge and S3 🙌
 
 ## 🤔 How it works
 
-You can read the [EventBridge Message Bus Adapter documentation](https://www.npmjs.com/package/@castore/event-bridge-message-bus-adapter) for regular cases.
+You can read the [EventBridge Message Bus Adapter documentation](https://www.npmjs.com/package/@castore/message-bus-adapter-event-bridge) for regular cases.
 
 When an entry is oversized, its `Detail` is saved as a JSON object in the provided s3 bucket. It's key is a concatenation of the constructor `s3Prefix` option, the `eventStoreId` and `aggregateId` of the event and the current timestamp:
 
@@ -86,7 +86,7 @@ On the listeners side, you can use the `EventBridgeS3MessageBusMessage` TS type 
 import {
   EventBridgeS3MessageBusMessage,
   parseMessage,
-} from '@castore/event-bridge-s3-message-bus-adapter';
+} from '@castore/message-bus-adapter-event-bridge-s3';
 
 const listener = async (
   message: EventBridgeS3MessageBusMessage<typeof appMessageBus>,
@@ -104,7 +104,7 @@ import fetch from 'node-fetch';
 import {
   EventBridgeS3MessageBusMessage,
   parseMessage,
-} from '@castore/event-bridge-s3-message-bus-adapter';
+} from '@castore/message-bus-adapter-event-bridge-s3';
 
 const listener = async (
   message: EventBridgeS3MessageBusMessage<typeof appMessageBus>,
@@ -116,7 +116,7 @@ const listener = async (
 You can provide event store ids and event types if you listener only listens to specific event types:
 
 ```ts
-import type { EventBridgeS3MessageBusMessage } from '@castore/event-bridge-s3-message-bus-adapter';
+import type { EventBridgeS3MessageBusMessage } from '@castore/message-bus-adapter-event-bridge-s3';
 
 const listener = async (
   message: EventBridgeS3MessageBusMessage<
