+++ packages/message-bus-adapter-event-bridge-s3/babel.config.js
