+++ packages/message-bus-adapter-event-bridge-s3/dependency-cruiser.js
