+++ packages/message-bus-adapter-event-bridge-s3/src/parseMessage.ts
