+++ packages/message-bus-adapter-event-bridge-s3/tsconfig.json
