+++ packages/message-bus-adapter-event-bridge-s3/vite.config.js
