+++ packages/message-bus-adapter-event-bridge/.eslintignore
