+++ packages/message-bus-adapter-event-bridge/.lintstagedrc.js
