+++ packages/message-bus-adapter-event-bridge/dependency-cruiser.js
