+++ packages/message-bus-adapter-event-bridge/src/adapter.ts
