+++ packages/message-bus-adapter-event-bridge/src/message.ts
