+++ packages/message-bus-adapter-event-bridge/tsconfig.json
