+++ packages/message-bus-adapter-event-bridge/vite.config.js
