+++ packages/message-bus-adapter-in-memory/.eslintignore
