+++ packages/message-bus-adapter-in-memory/.lintstagedrc.js
