+++ packages/message-bus-adapter-in-memory/babel.config.js
