+++ packages/message-bus-adapter-in-memory/dependency-cruiser.js
