+++ packages/message-bus-adapter-in-memory/package.json
