+++ packages/message-bus-adapter-in-memory/project.json
