+++ packages/message-bus-adapter-in-memory/src/adapter.ts
