+++ packages/message-bus-adapter-in-memory/src/message.ts
