+++ packages/message-bus-adapter-in-memory/src/types.ts
