+++ packages/message-bus-adapter-in-memory/src/utils.ts
