+++ packages/message-bus-adapter-in-memory/tsconfig.build.json
