+++ packages/message-bus-adapter-in-memory/tsconfig.json
