+++ packages/message-bus-adapter-in-memory/vite.config.js
