+++ workspace.json
@@ -8,9 +8,9 @@
     "json-schema-event": "packages/json-schema-event",
     "zod-event": "packages/zod-event",
     "json-schema-command": "packages/json-schema-command",
-    "event-bridge-message-bus-adapter": "packages/event-bridge-message-bus-adapter",
-    "event-bridge-s3-message-bus-adapter": "packages/event-bridge-s3-message-bus-adapter",
-    "in-memory-message-bus-adapter": "packages/in-memory-message-bus-adapter",
+    "message-bus-adapter-event-bridge": "packages/message-bus-adapter-event-bridge",
+    "message-bus-adapter-event-bridge-s3": "packages/message-bus-adapter-event-bridge-s3",
+    "message-bus-adapter-in-memory": "packages/message-bus-adapter-in-memory",
     "sqs-message-queue-adapter": "packages/sqs-message-queue-adapter",
     "in-memory-message-queue-adapter": "packages/in-memory-message-queue-adapter",
     "test-tools": "packages/test-tools",
